t.rast.algebra/testsuite: split file test_raster_algebra.py (#2974)
In order to address the frequent CI timeout during execution of the `test_raster_algebra.py` file this file is split into two parts with roughly similar test duration.

Also rename split files.
